refactor(agents): Extract orderAgentIds/shuffleArray to shared utility
- Create packages/agents/src/shared/agent-ordering.ts with shared utilities
- Export AgentOrderingStrategy, OrderableAgent, orderAgentIds, shuffleArray
- Remove duplicate implementations from route.ts and test file
- Fix multi-word keyword matching in complexity detection
- Fix probabilistic test assertion (> instead of >=)
