+++ packages/agents/src/services/index.ts
@@ -28,6 +28,9 @@ export {
   type AgentOrderingStrategy,
   detectMessageComplexity,
   type MessageComplexity,
+  type OrderableAgent,
+  orderAgentIds,
+  shuffleArray,
   TeamChatResponseService,
   teamChatResponseService,
   type UntaggedResponseConfig,
