+++ packages/agents/src/shared/index.ts
@@ -4,5 +4,6 @@
  * @packageDocumentation
  */
 
+export * from './agent-ordering';
 export * from './snowflake';
 export * from './types';
